requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.1.8
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.49.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
